Paid order Delete from our plugin record
If the order # 1 has been considered as the abandoned and it is
displayed under the Abandoned orders tab. Then user will pay for the
order VIA PAYPAL.

Note: Once the Order is placed via paypal, mark the order status as
COMPLETED from admin end.

And user do not reached the Order received page. Then the same order
still marked as the abandoned, where as the order is completed because
it is paid the amount.

So for this record user was receiving the abandoned cart email
notifications.

So above fix ensure that this kind of order will be deleted from our
plugin so abandoned cart notification will not send. It will apply for
guest and the logged in user.

Also the order status PENDING and FAILED will be consider as the
abandoned and the abandoned cart notification will be send for that
orders.