Changes are listed below in description
1) Settings API has been implemented for the ac lite settings.

If new install then default setting cut off time will be 60, and it is
updated via ftp then it will carry forward the old settings.

2) Default template will be installed when plugin is first time
installed.
It will be not active, so user can check the content of the template.

3) New setting added in the add / edit template: Use WooCommerce
Template Style : It will use the WC templates Headers & footer for the
abandoned cart email notifications.

This setting will be enabled for default template and if we add new
template than it will be not enable.

4) The Use WooCommerce Template Style :  setting contain the test view
of the template with the WC settings and without WC settings.

5) New short code has been introduced "{{cart.abandoned_date}}" it will
print the date and time when the cart was abandoned.