Fix #16
It is manly for the Logged in user: When User A have abandoned the cart
we used to send the email to the Email address which was used while
creating the account on respective site.

But, the abandoned email should be send to the email address which was
entered on the checkout page . This Is not the on change event so if
user have updated his profile biling address then it will be takes, it
is not like the We are stading on the checkout page and we have changed
the billing address but did not updated the billing address.

So this fix ensure that the abandoned cart email will be sent to the
email address which was entered on the checkout page.

This fix also ensure that the same email address will be displayed on
the Abandoned orders list page