Version 2.5
Abandoned cart Lite version 2.5 Changelog

There are a few changes made in this version which are mentioned as
below:

1. The Settings page for the plugin has been redone. We are now using
the WordPress Settings API for all the options on that page.

2. When the plugin is installed & activated for the first time, 1
default email template will also be created. The template will be
inactive when the plugin is activated.

3. A new setting is added on the Add/Edit Template page named as
“Active”. If this setting is checked, only then the abandoned cart
reminder email will be sent to the customers. If this setting is
unchecked, then the email template won't be sent to the customers, but
still you can keep it in the plugin. By default, this is unchecked.

4. A new setting is added on the Add/Edit Template page named as “Use
WooCommerce Template Style”. If this setting is checked then abandoned
cart reminder email will use the WooCommerce style (header, footer,
background, etc.) for the notifications. If it is not checked then the
regular email will be sent to the customer as per the formatting that is
set in the template editor.

For existing users, this setting will remain unchecked. For new users of
the plugin, the setting will be enabled for the existing default email
template that is provided with the plugin.

5. Abandoned cart email notification will be sent to the client's
billing address entered on checkout instead of on the email address
added by the user while registering to the website. This applies only
for logged in users.

6. New shortcode “{{cart.abandoned_date}}” has been introduced in this
version. It will display the date and time when the cart was abandoned
in the abandoned cart email notification.

7. When a customer places an order within the abandoned cart cut off
time, then the order received page was displaying a warning. This has
been fixed.

8. Abandoned Orders tab was not sorting according to the “Date” column.
Same way, Recovered Orders tab was not sorting according to “Created On”
& “Recovered Date” column. This has been fixed.

9. Some warnings were displayed on the Abandoned Orders, Recovered
Orders and Product Report tab. These have been fixed.

10. The 'mailto' link was not working on the Abandoned Order details
page. This has been fixed.

Tweak:

1. Removed the background white color for the add / edit template page.

2. Abandoned Orders tab will display the user’s billing address using
which the cart was abandoned. This applies only for logged in users.