Changes are listed below
1) When user changes the WordPress directory names then our plugin do
not allow to send the emails, because it was not able to include the
"wp-load.php" file in the "send_email.php" file.

User was changing the directory name via plugin. So this fix ensure that
if "WP_CONTENT_DIR" is defined then use the constant name and then
include the wp-load.php file.

2) If the cart is empty and we have tried to place the order via email
then recover date was displaying the wrong date of recovery order on
Recover Order tab.

It has been fixed. This issue was raised by the client.