New settings field has been added to template
Two new settings fields has been added in Add / edit template.
1) Send From This Email Address:
2) Send Reply Emails to:

This email ids will be added in the header of the TEST email as well as
the abandoned cart email notifications.

So once user click on the reply to the abandoned cart notification it
will automatically the email address of the headers.

Default value will be the WordPress- setting- email id.