Changes are listed Below. Fix #29, Fix #28
This commit have 2 new settings for the add / edit template page.
1) Send From This Email Address: It will be appeared when normal email
template is send to the use, It should not have WC template style.
2) Send Reply Emails to:

Both address will be displayed in the reminder emails Headers part

fix 29 :  We have user meta named as persistent cart which store the
abandoned cart data. So in lite version we have not deleted the entry
while deleting the plugin,
Due to which while testing it was taking the old user meta and
displaying wrong product.

Fix 28 : Our plugin From name and Send From This Email Address: was
overwritten by the WC setting , so we have given notice for it.