Version 2.6
The change log for this is as a below:

There are 2 new features and 1 issue fix in this version which are
mentioned as below:

1) New setting named as “Send From This Email Address:” is added in Add
/ Edit template page. It will allow to change the From Email address of
abandoned cart notifications.

2) New setting named as “Send Reply Emails to:” is added in Add / Edit
template page. It will allow to change the Reply to email address of the
abandoned cart notifications.

3) If the wp-content folder name is changed using iThemes Security
(formerly Better WP Security) plugin then abandoned cart email
notification were not sent. This has been fixed.