Abandoned orders Blank cart
When logged in user / guest user make the cart empty before the
abandoned cart cut-off time is reached the that cart entry becomes the
blank in the db. And the same cart is displayed in the abandoned orders
list page.

This fix ensure that the blank cart of the logged in users as well as
the guest user will not be displayed.