Version 2.7
Change log for Version 2.7

1. New setting named as “Email Template Header Text:” is added in Add /
Edit template page. It will allow to change the header text of the email
which have WooCommerce template style setting enabled for the template.

2. From this version, the email sending process will run every 15
minutes instead of every 5 minutes. This will result in improved overall
performance of the website.

3. When Lite version of the plugin is activated on the site then it was
not allowing to activate the PRO version of the plugin. This has been
fixed.

4. When templates are created / updated and if it has the same duration
as one of the existing templates, then new template was not saved. This
has been fixed.