Fix #47 , Fix #48
We have one error displayed in the console which was indicating that it
s because  of the jQueryui. So We have  upgraded the script version to
the latest version which is 1.11.4.

After this the default template was displaying the table content in the
visual mode when we first time load the page.

Also, newly added template in which we add the content in the VISUAL
mode was not adding the content, this is also fixed.