Test emails will not send when email body content will blank.
When email body content will blank then our plugin will not sent the
test email and 'Email has been not sent. There are no proper Email
content.' message will be display.

We were getting the email body content from the TinyMCE editor which was
not allowed to send the test email from text mode of the tinyMCE.

This issue has been fixed. Now we are getting the content from the
specific editor to send the test emails. It will allow sending the test
email from the "Visual" mode and "Text" mode of the email body.