Changes are listed below:
1)  On click of the product image of the short code: {{products.cart}}
Will redirect to CART page. Earlier it was redirecting to product page.

2) Logic for including the wp-load has been changed. So please carefully
check plugin work with the fresh install and update.

Note: Also check by changing the plugin directory name of wp-content.

3) added new encryption algorithm:

We have added same Pro version's algorithm for encryption the data.

Note: Old email should be work with this new copy. So send 3-4 emails
before updating the plugin.

4) Number format to ROUND method: Earlier we had used number format.
So check by providing the 1000+ total of the cart, and decimal values
also.

5) On abandoned order tab we have now displayed billing address. If
billing address is not set then show the users email address which was
used while registration.

6) Guest cart capturing logic changed. Implemented the same Pro version
logic for capturing the guest carts. Here also check by products which
have images.

7) We have now started the session on required pages. This is also
implemented from the Pro version. But in lite we have session while we
open the cart from the email.

Note: Check specially for the guest carts.