Blank Row displayed for Guest
We have an issue where it was displaying the Blank Row at the abandoned
orders tab.

We have fixed the same in pro version.

Same need to be check in this version.

Product must have the description and in that description html tags ,
images , special character.

It should capture the GUEST CART and should display it.