Version 2.8
Change log for 2.8 is as below:

We have done some major modifications to the following components:
Encryption & Performance

Bug fixes:

1. We have changed the encryption for the links that are sent in the
Abandoned cart email notifications. Earlier we were using the mcrypt
library for encoding the data. If mcrypt library was not installed on
the server, then abandoned cart email notifications were not sent. Now
we have used different functions for encoding the string. We have used
microtime function & a security key. Using this security key, and after
applying an algorithm on it, we generate the encoded string.

2. The session now starts only on required pages of the store. Earlier
it was started globally. This will help to improve the site performance.

3. If billing email address of the logged-in user is not set then it was
showing blank space on the abandoned orders list. This has been fixed.
Now it will show the email address which was used while registering to
the store.

4. Earlier if email body was blank and we send the test email then blank
email was sent. This has been fixed. Now if email body is blank then
test email will not be sent.

Tweaks:

1.   Earlier we were populating the guest cart information by looping
into the global WooCommerce cart. Now we are not looping & instead using
the WooCommerce session object itself.

2. Earlier if the ‘wp-content’ directory has been renamed, then
wp-load.php file was not included and abandoned cart reminder email was
not sent.  Now, we have changed the way of including the wp-load.php
file for sending the abandoned cart reminder notifications.

3. Earlier when {{products.cart}} merge tag is used in abandoned cart
email template, then on click of the product name and product image, it
was redirecting to the product page. Now it will redirect the user to
the cart page.

4. We are now rounding off the prices with the “round” function.