Fix #66
When there is no abandoned order under the Abandoned Order tab and we
add the product to the cart but cut off time is not reached then it was
showing checkbox and a total number of items were increasing. And we
were not shown "No items found" message.

This fix ensures that Abandoned Order tab will display the message "No
items found" and items will not increase until cut off time reached.

This issue has been fixed in the PRO version.