Recovered logic & Unsubscribe Short code and link
Recoved order logic:

Earlier if any user came from abandoned cart reminder email and place
the order using PayPal payment gateway and do not reach the order
received page. Then plugin was not considering that order as a
recovered. From now onwards if the user came from the abandoned cart
reminder email and place the order using PayPal and do not reach the
order received the page. Then plugin will consider that cart as a
recovered order.

This the same logic as we we have applied in the pro version.

Unsubscribe :

We have added new short code / merge code in email template.

This {{cart.unsubscribe}} merge code will be replaced with unsubscribe
link in the reminder email email.

It has the same functionality as the PRO version.