New setting has been added
We have added new setting named as "Display Taxes separately" under the
Setting tab.

Earlier we included the tax into the product price and then calculate
the Cart Total.

Now, if this setting is checked then new separate row named as "Tax"
will be included in the abandoned cart reminder email and then we will
display the "Cart Total" at the end. It will display only price of the
product under the column of the price and then it will show the total
tax of the product price in the new row "Tax" and then we will calculate
the cart total with tax.

And if this setting will unchecked then taxes will be added to the
product price & no separate row for Tax will be shown in the abandoned
cart email notification.

We have also added preview email template for this setting.

The same scenario will be considered for the test email and the setting
"Use WooCommerce Template Style" in the Add / Edit email template.