Fix #70
When we place an order after cut-off time but emails are not sent then
recovered cart was displaying under the "Recovered Order" tab.

This fix ensure that the user will place an order after cut-off time but
emails are not sent then our plugin will not consider as a recovered
order.