Fix #71
The database error was displaying on the order received page. As we are
updating the unknown column on sent history table.

This fix ensure that when user will place an order from same browser and
window after getting the reminder email then the same order will
consider as recovered order. And the database error will not show on
order received page.