fix #72
The issue was with the Logged-in user.

When product name have the special character then it was showing the
blank cart on the abandoned orders page.

This is because we have added function addslashes in the cart info. So
it was adding the slashes all around the string which was wrong. It
shuold only add the slashes in the product name.

So this fix ensure that for a Logged - in user if cart is abandoned with
the special character in the product name.