Added unsubscribe link in test email and default template
We have added unsubscribe link in the test email and default email
template.

In the test email, unsubscribe link will be clickable but it does not
redirect to another page.

When we will first time install the plugin, the unsubscribe link will be
included in the default email template of Email Body.