Fix #68
When the user place an order from same browser after cut-off time is
reached and emails are sent then the record was not coming under the
recover order tab.As we the record is deleted from the abandoned history
table.

This issue has been fixed. Now the user will place an order from the
same browser and window after getting emails then the order will display
under the "Recovered Order" tab.