Version 2.9
Abandoned cart Lite version 2.9 Change log

This version has changes in the way recovered orders are captured along
with one new feature.

New Feature:
1. New merge tag {{cart.unsubscribe}} has been added to add / edit
template. This merge tag allows user to stop receiving further abandoned
cart reminder email notifications. This merge tag has been added to
comply with email sending laws in different countries.

Bugs Fixed:
1. Earlier if any user came from abandoned cart reminder email and place
the order using PayPal payment gateway and do not reach the order
received page. Then plugin was not considering that order as a recovered
order. From now onwards if the user came from the abandoned cart
reminder email and place the order using PayPal and does not reach the
order received the page. Then plugin will consider that cart as a
recovered order.
2. When the cart is abandoned as a guest & product have the special
character in the attributes name, then it was displaying a blank row
with only a checkbox on the Abandoned Orders tab. This has been fixed.

Tweak:

1. If the order is recovered from the abandoned cart reminder email then
it will add a note “This order was abandoned & subsequently recovered.”
for the order.