Fix #85
We are using the site URL to load the jQuery on the checkout page. On
the Client's site the site URL was wrong so jQuery was not properly
loaded, due to which abandoned cart was not captured.

Now we will use the admin URL to load the jQuery on the checkout page.
It will consider the path (<site url>/wp-admin ) and allow to capture
the abandoned carts.