Fix #87
The Recovered Orders was displaying the duplicate record for single
order as we are updating the meta key before emails are not sent. When
the emails are sent and user place an order then we are again updating
the meta key.

This fix ensure that when user place an order after emails are sent then
recovered order tab will not update the order details to the last
recovered order.