Fix #89
The issue scenario is as below.
We have "Email admin On Order Recovery" setting enabled & we place the
order from the abandoned cart reminder email.

Then while sending the recovered cart reminder e mail to admin our
plugin was deleting the user meta. The user meta represent that the
order was abandoned, it is used to identify the abandoned order or the
USERS.

So once the ADMIN recovered email sent the user meta deleted, and as
user meta not found while registering the cart as an recovered it was
not displaying the cart as an recovered.

PLS : Check this fix With PayPal scenario where we do not reach the
Order received page. The admin Recovered email should come only Once.