fix #88
We have the blank row in which we do not have the abandoned order id,
date and the status. So while we sort that data it was showing the
error.

This fix will display the abandoned order id, date on which it was
abandoned, and the status.

This fix also allow to delete the blank row rom the tab.