fix #91
The issue occurred due to the billing email. We have considered the
billing email address for sending the abandoned cart reminder emails.

So if one order is placed for the day then we are deleting the another
order of the same users which have the same email address. So while
deleting the abandoned cart record we were still checking for the user
email address, not the billing email address.

So in this fix I have considered the billing email address for deleting
the new order of the same user on the same day.