fix #89
This fix ensure that the recovered carts will be displayed in the
recovered orders tab.

While recovered order email notifications is sent to the admin, we were
deleting the identifier for the abandoned carts. Once the identifier is
deleted the cart is not consider as the recovered.

Earlier if we do not delete that identifier then 2 times recovered cart
email notifications is sent to the admin.

To overcome the issue we have set one value in the post meta table with
the order id. So we can cross verify that the recovered email is sent
for the order id.