Fix #98
Our plugin was not creating the default email template on client's site
while activating the plugin.

Because we are using the "Innodb" as the engine type to create the
plugin's tables. The "Innodb" engine type was creating the issue and
tables were not created in the database.

To fix the issue, we have removed the engine type from queries and we
have used the WordPress collation to create the table.

This issue has been fixed in the PRO version.