Fix #99
Sometimes, the server does not have mcrypt_encrypt, MCRYPT_RIJNDAEL_256,
md5, & MCRYPT_MODE_CBC library enabled on it.

We are using the
encryption method in our plugin for sending the some information in the
URL, so to protect that information from others we encrypt our data.

The fix ensures that it will not require all above library and the data
will be secure, so no one can decrypt it. And when user will go cart
page through abandoned cart reminder email, it will not show blank page.