Changed the logic of Email Template
Let's say if our plugin had captured the 100 abandoned cart's in one
week but the email has been not sent to any of the users. And there is
total 3 Active templates on your store. Due WP Cron or any other issues,
the email has been not sent out to any of the 100 abandoned carts.

But, once the WP Cron issue or any other issue has been fixed then all
100 carts have reached the all 3 email templates time to sent the
reminder email. So in one time all 3 email templates will be sent to the
all 100 abandoned carts.

So in this case now we have checked that the first reminder email sent
time, And then we check that the next template time is greater than the
first sent email time.

If it is less then the first email sent time, then it will not send the
2nd email template.

E.x First reminder email template time is : 5 min.
second reminder email time is : 25 min.

Now I had abandoned cart at 11 am. And my both templates are
deactivated.
At 12 I have activated both the templates. So both email template time
has been reached.

So now first email template will be sent at 12.05 pm.

Now when we send the 2nd email template it will check 12.05 + 25 min (
template time ) And compare with the current time which is 12.05 where
we sent first email template.

So the current time is less then it will not send the 2nd email template

Note:

Please check above scenario with the normal reminder email flow. It
means that all cart reminder email should be sent as it was sending
earlier.