Use billing first and last name for logged-in users
If logged-in user abandoned the cart then we were used the First Name
and Last Name while creating the account. But the abandoned cart details
should be billing details of the checkout page.

This fix ensures that if logged-in user abandoned a cart then we will
display billing first name and billing last name in the Abandoned Order
list.

If logged in user had not set billing first name and last name then it
will consider the first name and last name which was used for creating
an account.

For the registered user, the merge code of the customer first name and
customer last name will replace with the billing first name and billing
last name respectively in the abandoned cart reminder email.

You need to check the Abandoned Orders tab, Abandoned Order details page
and Abandoned Cart Reminder email.