Manual order fix for recovery email
When cart is abandoned by the admin and then cart has been modified and
again it is abandoned. After that admin creating the new order for any
of the user then abandoned order recovery email is sent to the admin.

This fix ensure that the order which are placed from the CHECKOUT page
that order will consider for sending the recovery email to the admin.