Fix #108, Fix #54, Fix #90
This fix ensures that if any store has the wrong guest user id then we
will change the id for the guest user.

Fix #108: The email address was not populated on checkout page as the
guest user ID was wrong. The ID was starting with 1,2,3,... for the
guest users.

Fix #54 and Fix #90: Guest cart was coming wrong and empty cart through
abandoned cart reminder email on WOO_ACLITE as the guest user Id was
wrong.