Fix #115
I have noticed in the Loco translation plugin that it has wrong text
domain as "woocommerce-abandon-cart" for our plugin. And our plugin text
domain is "woocommerce-ac". So the plugin was not showing all strings
for translation.

I have changed the text domain as "woocommerce-ac". But it was showing
only half of strings for translation.

So I have created new ".POT" file, which will have all the translation
strings. And we could translate all the string of our plugin from the
Loco translation plugin.

Note:

You need to set the path for the "messages.pot" in the "Template file"
setting in "Advance" menu option.
Please see the screenshot for it: http://screencast.com/t/1o3AghIIhQi0