Fix #117
The client of PRO version had raised an issue that a large amount of
queries was running from our plugin. We have applied the same fix in the
LITE version. To prevent this issue we have removed unnecessary code
from the plugin file. And for that we have changed the order placed
after cut off time logic.

If the client is placed an order and for the
same day the client comes to the site and again try to abandon a cart
using the same email address then we were not capturing the cart.

Now,
for above scenario, we will capture the cart but after the cron time the
status of the cart will be changed to “Abandoned but new cart created
after this” on Abandoned Order tab. And Abandoned Cart reminder email
should not be sent.

Also, you need to check the abandoned cart reminder
email should send for the “Pending” and “Failed” order status.

Note:You
need to test for guest user and logged-in user.