Version 3.4
Change log for Version 3.4

This version has 3 enhancements.

Enhancements:

1. We have made our plugin compatible with the “Loco Translation”
plugin. Now, all the strings from Abandoned Cart Pro for WooCommerce
plugin can be translated using the Loco Translation plugin.

2. The email template are now compatible with the “WP Better Emails”
plugin. If the “WP Better Emails” plugin is activated on the site then
CSS style of this plugin will be applied to the test email and abandoned
cart reminder email.

3. When the email sending script is executed on the server, it was
running a large number of MySQL queries for guest users. That was
causing an increase in the server load. In this version, we have
optimised the queries.

Now the plugin will not fetch all the guest users record as it will
fetch the necessary record which needs to be checked. This will improve
the site performance.