Fix #131
We have added new feature for capturing visitor carts. For that, when we
abandoned a cart, it was capturing two carts- visitor and guest user.

This issue has been fixed. Now if we will abandon a cart then the
visitor's cart will be updated and it will show one cart on the
abandoned orders page.