Fix #166
When a guest user pays for an order using PayPal and the check for the
'create an account' is checked or enter the password on the checkout
page then that guest user becomes the Logged in user.

As the guest user becomes the registered user, the captured cart for the
guest was still in the database.

So the abandoned cart reminder email was sent to that guest cart.

Note: Here the order is placed directly from the browser. Customer is
not coming from the abandoned cart reminder email.

So this fix ensure that the guest cart and the logged in users cart will
be deleted and do not displayed under the abandoned orders list