changed in the last commit
I have changed the code of abandoned cart date and format which I had fixed in the last commit. I had used date_i18() function for translating abandoned cart date and time into WordPress site Language which was not allowed to insert the string between date and time. This function is used for retrieving the date in the localized format, based on timestamp.

This fix ensures that if the user wants to add the string between date and time, then the user can update the plugin file.