Fix #174, Fix #177, Fix #181
When the user placed an order and then again come to the site and abandoned a cart for the same day using the same email address then the new cart was being captured. This issue has been fixed.

Now, if the user abandoned a cart for the same day then our plugin will capture the cart but the cart status will be changed to 'Abandoned but new cart created after this' for the same cart after cron job run.