Added Tracking code
In this commi we have added the tracking code for the plguin. It will fetch the data of the clinet site according to the action taken by the users.

users will have 3 choices 1. allow 2. no thanks and 3.disclose

if user has allwed then it will capture the data from the site. if no and disclose the only abndoned and reocovred total amont will be fetched