Admin notices for pro version
We have added 4 notices at the admin side. It will display after 15 days, 22 days, 30 days after the first notice has been ignored

If The AC pro is activated the it will not show any notices.

In third notice that is 22 days, we are showing it for ORDD lite, so ordd pro is activated then we wil not display the same notice.