Fix #211
When three or more attributes were added for the variatble product, the
slug name of variation were showing instead of variation name on
Abandoned Orders details page and in the Abandoned Cart reminder email.

This issue has been fixed. Now, If number of attributes or number of
variable will be added for variable product then the variation name will
show on Abandoned Order details page and in the abandoned cart reminder
email.