Fix #260
When the guest user become the registered user then one blank abandoned
cart information was added in the database.

Now, our plugin will not be disaplyed blank cart details on the
Abandoned Orders tab. And the count for registerd user and for all user
under "All" tab will not consider blank cart.