Fix #67
When 'WP Editor' plugin is activated, then our template page was not
displayed.

The template page was blank. With the latest version of WooCommerce and
WordPress, I am unable to replicate the issue. But I have changes id
name of content which caused an issue.

So, Now, email template will not be blank when the WP Editor plugin is
activated.