Comply with GDPR
1) Added  2 settings for guest users and logged-in users to displayed
the message when tracking their carts under General setting tab.

If the setting will be blank then we do not show the notice and data
will be tracked.

If Admin adds the notice then we will show the notice on Email Address
field for guest users and on every Add to Cart button  for logged-in
users.

2) Added the privacy exporter and eraser files.

3) These will be called when personal data is exported or erased from WP
Dashboard->Tools->Export Personal data & WP Dashboard->Tools->Erase
Personal Data

4) Personal data will be erased based on the WooCommerce setting
Woocommerce->Settings->Accounts & Privacy->Remove personal data from
orders is set to on