Version 4.9
Changelog for Abandoned Cart LITE version 4.9

The plugin is now GDPR compliant.1. Added two settings to show the
message on the checkout page for guest users and the message to be
displayed at the start of the shop page for Registered users.2.
Abandoned Cart details will now be exported to the User’s Personal Data
from the Tools -> Export Personal Data menu and it can be erased from
Tools -> Erase Personal Data.3. Some notices are fixed in the debug.log
file.