Added new setting for the deletion of Abandoned Carts.
I have added new feature in the plugin. We will allow to customers to
delete the carts automatically after number of days.

We have added "Automatically Delete Abandoned Orders after X days" in
the "Settings" tab.

This will delete the abandoned carts from Abandoned Orders tab after the
number of days which is set in the setting.

It will also delete the record from database for sent emails.