Added New Default Email Template
Once the plugin will be updated, the new default email template will be
created in the list of the templates. It will be deactivated.

And once we installed the plugin, plugin was created one default email
template, it was simple. But we have changed it to attractive email
template.

Please test the test email, preview email templates, abandoned cart
reminder email with the new created email template.