Fix #384
One possible scenario where this issue can occur (the same cart is
created for a user after an order has already been placed) is if a guest
user becomes a registered user when placing the order.

This scenario was recently discovered in the PRO version. We will apply
a similar fix for the LITE version for the same scenario.

Issue:
https://github.com/TycheSoftwares/woocommerce-abandoned-cart-pro/issues/1703
Commit:
https://github.com/TycheSoftwares/woocommerce-abandoned-cart-pro/commit/84dc55eb98b07b35c3cdbffece1e755d0dc0d96f

Similar tests should be performed to ensure things continue to work
smoothly.