Added 2 new fields for GDPR
Until now, we've only shown 1 message for GDPR:
1. Guest Users - on the Checkout page
2. Registered Users - on the Shop & product page.

This has now been modified as below:
1. Guest Users - the message will be shown on the Checkout page
2. Registered Users - the message will be shown only on the product
page.

We will now provide the end user the ability to opt out of cart
tracking. To achieve this we've added 2 new settings:
1. opt out text
2. message to display when the user chooses to opt out

The Opt out text will be concatenated to the GDPR message when  it is
displayed. If no text is available, nothing will be displayed.

The text will be displayed as a link, when the user clicks on it, he/she
will see the message displayed in the Opt out message. If left blanks,
nothing will be displayed.

When the link is clicked, the AC record will be deleted (applies to all
users: visitors, guests, registered users)
For registered users, this will be set in the user meta table, so the
next time they log in, the GDPR message will not be displayed.