Change the default loaded tab when WC->Abandoned Orders menu is accessed
In the last update, a new tab Dashboard was added for the plugin which
is now loaded when WC->Abandoned Orders menu is accessed.

More than 1 client wants the Abandoned Orders tab to be loaded by
default.

So we've now added a hook using which the site admin can choose which
tab should be loaded.

Custom code which needs to be added to the functions.php file:

add_filter( 'wcal_default_tab', 'change_default_tab', 10, 1 );

function change_default_tab( $action ) {
return 'listcart';
}

In a scenario where the site admin wishes to load some other tab. the
tab name can be found by accessing the tab and checking the 'action'
value in the url.