Fix #641, Fix #642
The issue occurs as we fetch all the records, loop through them all and then create chunks based on the number of records to be displayed per page. So, when thousands of records are present, a memory allocation error occurs. Instead the code has now been modified to fetch the data, create chunks and process only the ones which we want to display.