Update code for 5.8.2
From this version on, we will be marking carts as complete in the cart history table when they finish looping through all the reminder email sequences. As a part of this, we have also marked all existing old carts as complete. This is being done specifically to improve the performance of scheduled actions.