Remove wp-load.php include.
Remove the include for wp-load.php. This is needed to load WP when the file is being executed in a manual cron job. Since the function to send emails is not being called anywhere except for the action scheduler function, it is not possible to execute the file in a manual cron job anymore. Hence removing the include for wp-load.php as it is not advisable to load WP as such for security reasons.