Pro settings updated & dashboard graph.
Added the new settings which were present in Pro to be displayed as a part of the Lite version. All these settings will be disabled by default. Users will need to purchase the Pro version to access these settings. The dashboard graph (a part of the Pro version) has now been included in the Lite version. The graph will display abandoned & recovered cart data based on the date range selected. By default, the current month data will be displayed.