Fix #737
An error is logged in debug.log when we try to send the test email for the default template. The error points to Emogrifier library. This is used by WC for inline css in emails. Updating the library should most likely fix the issue. Until then, I have removed some CSS which was causing the error. The default template should be checked to ensure it is displayed fine across different screen sizes such as mobiles, tablets and laptops.